requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0,<1.0
uvicorn[standard]>=0.29.0,<1.0
python-multipart>=0.0.9,<1.0
jinja2>=3.1.4,<4.0

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1,<2.0

# ── LangChain ─────────────────────────────────────────────────────────────────
langchain>=0.2.0,<1.0
langchain-community>=0.2.0,<1.0
langchain-text-splitters>=0.2.0,<1.0

# ── Vector store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0,<2.0

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0,<4.0

# ── PDF loader ────────────────────────────────────────────────────────────────
pypdf>=4.0.0,<5.0

# ── HTTP client (Groq API) ────────────────────────────────────────────────────
requests>=2.32.0,<3.0
